refactor: move Maven utility to a new package
