+++ internal/utility/maven/fixtures/my-app/pom.xml
