+++ internal/utility/maven/fixtures/parent/pom.xml
@@ -1,0 +1,8 @@
+<!-- For testing parent relative path -->
+<project>
+
+  <groupId>org.test</groupId>
+  <artifactId>parent-pom</artifactId>
+  <version>1.0.0</version>
+
+</project>
