+++ internal/utility/maven/fixtures/pom.xml
